+++ cluster/bootstrap.go
@@ -35,7 +35,6 @@ type AddressProvider interface {
 // Bootstrapper performs a bootstrap of this node.
 type Bootstrapper struct {
 	provider  AddressProvider
-	expect    int
 	tlsConfig *tls.Config
 
 	joiner *Joiner
@@ -48,10 +47,9 @@ type Bootstrapper struct {
 }
 
 // NewBootstrapper returns an instance of a Bootstrapper.
-func NewBootstrapper(p AddressProvider, expect int, tlsConfig *tls.Config) *Bootstrapper {
+func NewBootstrapper(p AddressProvider, tlsConfig *tls.Config) *Bootstrapper {
 	bs := &Bootstrapper{
 		provider:  p,
-		expect:    expect,
 		tlsConfig: &tls.Config{InsecureSkipVerify: true},
 		joiner:    NewJoiner("", 1, 0, tlsConfig),
 		logger:    log.New(os.Stderr, "[cluster-bootstrap] ", log.LstdFlags),
@@ -87,7 +85,6 @@ func (b *Bootstrapper) Boot(id, raftAddr string, done func() bool, timeout time.
 	tickerT := time.NewTimer(jitter(time.Millisecond))
 	defer tickerT.Stop()
 
-	notifySuccess := false
 	for {
 		select {
 		case <-timeoutT.C:
@@ -104,27 +101,31 @@ func (b *Bootstrapper) Boot(id, raftAddr string, done func() bool, timeout time.
 			if err != nil {
 				b.logger.Printf("provider lookup failed %s", err.Error())
 			}
-			if len(targets) < b.expect {
+
+			if len(targets) == 0 {
 				continue
 			}
 
-			// Try an explicit join.
+			// Try an explicit join first. Joining an existing cluster is always given priority
+			// over trying to form a new cluster.
 			b.joiner.SetBasicAuth(b.username, b.password)
 			if j, err := b.joiner.Do(targets, id, raftAddr, true); err == nil {
 				b.logger.Printf("succeeded directly joining cluster via node at %s", j)
 				return nil
 			}
 
-			// Join didn't work, so perhaps perform a notify if we haven't done
-			// one yet.
-			if !notifySuccess {
-				if err := b.notify(targets, id, raftAddr); err != nil {
-					b.logger.Printf("failed to notify all targets: %s (%s, will retry)", targets,
-						err.Error())
-				} else {
-					b.logger.Printf("succeeded notifying all targets: %s", targets)
-					notifySuccess = true
-				}
+			// This is where we have to be careful. This node failed to join with any node
+			// in the targets list. This could be because none of the nodes are contactable,
+			// or none of the nodes are in a functioning cluster with a leader. That means that
+			// this node could be part of a set nodes that are bootstrapping to form a cluster
+			// de novo. For that to happen it needs to now let the otehr nodes know it is here.
+			// If this is a new cluster, some node will then reach the bootstrap-expect value,
+			// form the cluster, beating all other nodes to it.
+			if err := b.notify(targets, id, raftAddr); err != nil {
+				b.logger.Printf("failed to notify all targets: %s (%s, will retry)", targets,
+					err.Error())
+			} else {
+				b.logger.Printf("succeeded notifying all targets: %s", targets)
 			}
 		}
 	}